Added ability to edit education to users page
