+++ static/js/containers/UserPage_test.js
@@ -15,7 +15,10 @@ import {
 } from '../actions';
 import {
   SET_WORK_DIALOG_VISIBILITY,
-  SET_WORK_DIALOG_INDEX
+  SET_WORK_DIALOG_INDEX,
+  SET_EDUCATION_DEGREE_LEVEL,
+  SET_EDUCATION_DIALOG_INDEX,
+  SET_EDUCATION_DIALOG_VISIBILITY,
 } from '../actions/ui';
 import IntegrationTestHelper from '../util/integration_test_helper';
 import * as api from '../util/api';
@@ -39,6 +42,81 @@ describe("UserPage", () => {
 
   afterEach(() => {
     helper.cleanup();
+  });
+
+  describe("Education History", () => {
+    it('shows the education component', done => {
+      renderComponent(`/users/${SETTINGS.username}`, userActions).then(([, div]) => {
+        let title = div.getElementsByClassName('profile-card-title')[1];
+        assert.equal(title.textContent, 'High school');
+        done();
+      });
+    });
+
+    it('should let you delete an education entry', done => {
+      renderComponent(`/users/${SETTINGS.username}`, userActions).then(([, div]) => {
+        let updatedProfile = _.cloneDeep(USER_PROFILE_RESPONSE);
+        updatedProfile.education.splice(0,1);
+
+        patchUserProfileStub.throws("Invalid arguments");
+        patchUserProfileStub.withArgs(SETTINGS.username, updatedProfile).returns(
+          Promise.resolve(updatedProfile)
+        );
+
+        let deleteButton = div.getElementsByClassName('profile-tab-card')[1].
+          getElementsByClassName('profile-row-icons')[0].
+          getElementsByClassName('mdl-button')[1];
+
+        listenForActions([
+          START_PROFILE_EDIT,
+          UPDATE_PROFILE_VALIDATION,
+          REQUEST_PATCH_USER_PROFILE,
+          RECEIVE_PATCH_USER_PROFILE_SUCCESS,
+        ], () => {
+          TestUtils.Simulate.click(deleteButton);
+        }).then(() => {
+          done();
+        });
+      });
+    });
+
+    it('should let you edit an education entry', done => {
+      renderComponent(`/users/${SETTINGS.username}`, userActions).then(([, div]) => {
+
+        let editButton = div.getElementsByClassName('profile-tab-card')[1].
+          getElementsByClassName('profile-row-icons')[0].
+          getElementsByClassName('mdl-button')[0];
+
+        listenForActions([
+          SET_EDUCATION_DIALOG_INDEX,
+          SET_EDUCATION_DIALOG_VISIBILITY,
+          SET_EDUCATION_DEGREE_LEVEL,
+        ], () => {
+          TestUtils.Simulate.click(editButton);
+        }).then(() => {
+          done();
+        });
+      });
+    });
+
+    it('should let you add an education entry', done => {
+      renderComponent(`/users/${SETTINGS.username}`, userActions).then(([, div]) => {
+        let editButton = div.getElementsByClassName('profile-tab-card')[1].
+          getElementsByClassName('profile-add-button')[0];
+
+        listenForActions([
+          START_PROFILE_EDIT,
+          UPDATE_PROFILE,
+          SET_EDUCATION_DIALOG_INDEX,
+          SET_EDUCATION_DIALOG_VISIBILITY,
+          SET_EDUCATION_DEGREE_LEVEL,
+        ], () => {
+          TestUtils.Simulate.click(editButton);
+        }).then(() => {
+          done();
+        });
+      });
+    });
   });
 
   describe("Employment History", () => {
