+++ static/scss/dashboard.scss
@@ -180,7 +180,7 @@ c.validation-wrapper {
     width: 100%;
     position: relative;
     overflow: visible;
-    margin-bottom: 20px;
+    margin-bottom: 35px;
     padding-bottom: 50px;
 
     .profile-tab-card-grid:first-child {
